+++ DataTable/DataTableBuilder.php
@@ -403,7 +403,7 @@ class DataTableBuilder implements ColumnFactoryAwareInterface
         return !empty($this->searchableKeys);
     }
 
-    private function sorting()
+    public function sorting()
     {
         if ($this->sortingColumn !== null) {
             $this->columnSorting($this->columns[$this->sortingColumn], $this->sortingOrder);
@@ -459,7 +459,6 @@ class DataTableBuilder implements ColumnFactoryAwareInterface
     private function prepareData()
     {
         $this->filtering();
-        $this->sorting();
     }
 
     /**